requirements.txt
tapo;
python-dotenv;